LayoutTests/accessibility/custom-elements/heading-expected.txt

This tests that aria fallback roles work correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.accessibleElementById("heading-1").intValue is 1
PASS accessibilityController.accessibleElementById("heading-1-1").intValue is 2
PASS successfullyParsed is true

TEST COMPLETE
1. Some title 1.1. Subtitle
